backend/requirements.txt
Django == 5.0.3
django-debug-toolbar == 4.3.0
djangorestframework == 3.14.0
django-cors-headers == 4.3.1
drf_social_oauth2==2.1.3
decouple==0.0.7

# djangorestframework-simplejwt == 5.3.1
# dj-rest-auth == 2.1.10
# djoser == 2.2.2
# PyJWT == 2.8.0
